django/requirements.txt
Django==2.1.7
django-cors-headers==2.5.2
psycopg2==2.7.7
postgis==1.0.4
djangorestframework==3.9.2
djangorestframework-gis==0.14
django-cors-headers==2.5.2
